Move community page components to the right place
