+++ src/components/pages/community/sections/discussions/DiscussionPlatforms.tsx
