+++ src/components/pages/community/shared/ContentCard/ContentCard.module.css
