+++ src/components/pages/community/shared/ContentCard/ContentCard.tsx
