+++ src/components/pages/community/shared/Section/Section.tsx
