+++ src/components/pages/community/shared/Slider/Slider.tsx
